+++ example/lib/main.dart
@@ -21,6 +21,7 @@ class MyApp extends StatefulWidget {
 
 class _MyAppState extends State<MyApp> {
   OpenResult? _openResult;
+  final _fileNameController = TextEditingController(text: '/sdcard/test.txt');
 
   Future<String> _downloadFile(String url, String filename) async {
     if (kIsWeb) {
@@ -56,12 +57,12 @@ class _MyAppState extends State<MyApp> {
         appBar: AppBar(
           title: const Text('Plugin example app'),
         ),
-        body: Center(
+        body: SingleChildScrollView(
           child: Column(
             mainAxisAlignment: MainAxisAlignment.center,
             children: [
               Padding(
-                padding: const EdgeInsets.all(8.0),
+                padding: const EdgeInsets.symmetric(horizontal: 8.0, vertical: 24.0),
                 child: Text(
                   _openResult == null
                       ? 'Result: none'
@@ -75,7 +76,9 @@ class _MyAppState extends State<MyApp> {
                   _openFile(await createTextFile(_generateRandomString(30)));
                 },
               ),
-              SizedBox(height: 8.0,),
+              SizedBox(
+                height: 8.0,
+              ),
               ElevatedButton(
                 child: Text('Download and open image'),
                 onPressed: () async {
@@ -83,7 +86,9 @@ class _MyAppState extends State<MyApp> {
                       'https://picsum.photos/200/300', 'test.jpg'));
                 },
               ),
-              SizedBox(height: 8.0,),
+              SizedBox(
+                height: 8.0,
+              ),
               ElevatedButton(
                   child: Text('Download and open calendar event'),
                   onPressed: () async {
@@ -91,11 +96,34 @@ class _MyAppState extends State<MyApp> {
                         'https://raw.githubusercontent.com/yendoplan/open_app_file/master/example/files/test.ics',
                         'test.ics'));
                   }),
-              SizedBox(height: 8.0,),
+              SizedBox(
+                height: 8.0,
+              ),
               ElevatedButton(
                 child: Text('Open non-existent file'),
-                onPressed: () async {
+                onPressed: () {
                   _openFile('asdf.qwert');
+                },
+              ),
+              Padding(
+                padding: const EdgeInsets.all(24.0),
+                child: Text('To test external file access on Android:\n'
+                    '1. Create or download a file to a restricted location (for example, "test.txt" in the root of external storage)\n'
+                    '2. Set the file path in the field below and try to open it\n'
+                    '3. Go to system settings and grant access to all files\n'
+                    '4. Try to open the same file again\n'
+                    'Use similar approach for media files: make an image like /sdcard/test.png and grant READ_EXTERNAL_STORAGE permission to the app'),
+              ),
+              Padding(
+                padding: const EdgeInsets.symmetric(horizontal: 24.0),
+                child: TextField(
+                  controller: _fileNameController,
+                ),
+              ),
+              ElevatedButton(
+                child: Text('Open file'),
+                onPressed: () async {
+                  _openFile(_fileNameController.value.text);
                 },
               ),
             ],